requirements.txt
openai==0.27.0
feedparser==6.0.8
requests==2.26.0
python-wordpress-xmlrpc==2.3
python-dotenv==0.19.1
